\name{site1}
\docType{data}
\alias{site1}
\title{Example individual-level data of data-contributing site 1}
\description{
This dataset gives an example of individual-level data of data-contributing site 1.
}
\usage{data(site1)}
\format{A data frame consisting of 800 individuals (i.e., rows) with variables (i.e., columns) sequence number (id), outcome (time), non-censoring indicator (status), exposure/treatment indicator (A), five baseline covariates (X1-X5), and data-contributing site indicator (indSite).}
\keyword{datasets}
