\name{site2}
\docType{data}
\alias{site2}
\title{Example individual-level data of data-contributing site 2}
\description{
This dataset gives an example of individual-level data of data-contributing site 2.
}
\usage{data(site2)}
\format{A data frame consisting of 1000 individuals (i.e.,rows) with variables (i.e., columns) sequence number (id), outcome (time), non-censoring indicator (status), exposure/treatment indicator (A), five baseline covariates (X1-X5), and data-contributing site indicator (indSite).}
\keyword{datasets}
