\name{site3}
\docType{data}
\alias{site3}
\title{Example individual-level data of data-contributing site 3}
\description{
This dataset gives an example of individual-level data of data-contributing site 3.
}
\usage{data(site3)}
\format{A data frame consisting of 1200 individuals (i.e.,rows) with variables (i.e., columns) sequence number (id), outcome (time), non-censoring indicator (status), exposure/treatment indicator (A), five baseline covariates (X1-X5), and data-contributing site indicator (indSite).}
\keyword{datasets}
